Chore: 코드 정리#89
