+++ MUMENT/MUMENT/Global/Extensions/UIImageView+.swift
@@ -40,21 +40,4 @@ extension UIImageView {
             }
         }
     }
-//
-//    extension UIImageView {
-//        func load(_ URLAddress: String) {
-//            guard let url = URL(string: URLAddress) else {
-//                return
-//            }
-//            DispatchQueue.global().async { [weak self] in
-//                if let data = try? Data(contentsOf: url) {
-//                    if let image = UIImage(data: data) {
-//                        DispatchQueue.main.async {
-//                            self?.image = image.resizeImage()
-//                        }
-//                    }
-//                }
-//            }
-//        }
-//    }
 }
