+++ MUMENT/MUMENT/Network/Services/SongDetailService.swift
@@ -7,17 +7,6 @@
 
 import Alamofire
 
-/*
- AuthRouter : 여러 Endpoint들을 갖고 있는 enum
- TargetType을 채택해서 path, method, header, parameter를 각 라우터에 맞게 request를 만든다.
- */
-
-//// MARK: - SongDetailBodyModel
-//struct SongDetailBodyModel: Codable {
-//    var musicId: String
-//    var userId: String
-//}
-
 enum SongDetailService {
     case getSongInfo(musicId: String, userId: String)
     case getAllMuments(musicId: String, userId: String, isOrderLiked: Bool)
